Update console operator and connectors to use latest pkg and separate gotohelm (#654)
* Update console operator and connectors to use latest pkg and separate gotohelm

* Fix up transpilation with gen

* Fix linter issues

(cherry picked from commit dc814fb6b3b8f04af56f8745c6c431995a4eb0fa)
