+++ charts/console/helpers.go
@@ -14,7 +14,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/redpanda-data/redpanda-operator/pkg/gotohelm/helmette"
+	"github.com/redpanda-data/redpanda-operator/gotohelm/helmette"
 )
 
 // Expand the name of the chart.
