+++ charts/console/notes.go
@@ -13,7 +13,7 @@ package console
 import (
 	"fmt"
 
-	"github.com/redpanda-data/redpanda-operator/pkg/gotohelm/helmette"
+	"github.com/redpanda-data/redpanda-operator/gotohelm/helmette"
 )
 
 func Notes(dot *helmette.Dot) []string {
