+++ pkg/lint/lint_test.go
@@ -194,7 +194,11 @@ func TestGoModLint(t *testing.T) {
 		// generator. The invocation `gen schema` uses reflection on Charts'
 		// Value types that need to be up to date. (Generation could be moved
 		// into their respective charts to resolve this).
-		modPrefix + "gen": {modPrefix + "charts/redpanda/v5", modPrefix + "charts/operator"},
+		modPrefix + "gen":                {modPrefix + "pkg", modPrefix + "charts/redpanda/v5", modPrefix + "charts/operator"},
+		modPrefix + "charts/operator":    {modPrefix + "pkg"},
+		modPrefix + "charts/redpanda/v5": {modPrefix + "pkg"},
+		modPrefix + "charts/connectors":  {modPrefix + "pkg"},
+		modPrefix + "charts/console":     {modPrefix + "pkg"},
 	}
 
 	// This could also be done with go work sync but go.work causes many other
